Fixed kpanic screen errors

+++ include/tupai/kpanic.hpp
@@ -27,7 +27,7 @@
 
 namespace tupai
 {
-	extern "C" void kpanic(const char* msg = nullptr, long code = 0x0) __attribute__((noreturn));
+	extern "C" void kpanic(const char* msg = nullptr, uint32 code = 0x0) __attribute__((noreturn));
 	extern "C" void kbreak();
 	extern "C" void khalt();
